+++ lib/pleroma/web/mastodon_api/controllers/media_controller.ex
@@ -26,7 +26,8 @@ defmodule Pleroma.Web.MastodonAPI.MediaController do
            ActivityPub.upload(
              file,
              actor: User.ap_id(user),
-             description: Map.get(data, :description)
+             description: Map.get(data, :description),
+             filename: Map.get(data, :filename)
            ) do
       attachment_data = Map.put(object.data, "id", object.id)
 
@@ -42,7 +43,8 @@ defmodule Pleroma.Web.MastodonAPI.MediaController do
            ActivityPub.upload(
              file,
              actor: User.ap_id(user),
-             description: Map.get(data, :description)
+             description: Map.get(data, :description),
+             filename: Map.get(data, :filename)
            ) do
       attachment_data = Map.put(object.data, "id", object.id)