+++ CHANGELOG.md
@@ -272,6 +272,7 @@ switched to a new configuration mechanism, however it was not officially removed
 - Mastodon API: Support for `bot` field in `/api/v1/accounts/update_credentials`.
 - Mastodon API: Support irreversible property for filters.
 - Mastodon API: Add pleroma.favicon field to accounts.
+- Mastodon API: Add `filename` parameter to `POST /api/v1/media` and `POST /api/v2/media`.
 - Admin API: endpoints for create/update/delete OAuth Apps.
 - Admin API: endpoint for status view.
 - OTP: Add command to reload emoji packs